requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
curl_cffi==0.15.0
cycler==0.12.1
dotenv==0.9.9
dukascopy-python==4.0.1
filelock==3.25.2
fonttools==4.62.1
frozendict==2.4.7
fsspec==2026.3.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
mt5==0.0.3.1
multitasking==0.0.12
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.29.7
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.6
plumbum==1.7.0
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PythonMetaTrader5==1.0.9
pytz==2026.1.post1
requests==2.33.1
rich==14.3.3
rpyc==5.2.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
xgboost==3.2.0
yfinance==1.2.1
